Moved process, window_handler and image_hanlder (newly added and empty) module to adapter folder
